requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
fastapi==0.90.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
isort==5.12.0
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycryptodome==3.17
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
starlette==0.23.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
